% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{get_required_field_names}
\alias{get_required_field_names}
\title{Get required field names for a data set. Possible fields are "Date" and "Close"}
\usage{
get_required_field_names()
}
\value{
List of field names.
}
\description{
Get required field names for a data set. Possible fields are "Date" and "Close"
}

