% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-series.R
\name{get_unmanaged_series}
\alias{get_unmanaged_series}
\title{Get unmanaged series from Quanl.}
\usage{
get_unmanaged_series(series_name, from_date)
}
\arguments{
\item{series_name}{String.}

\item{from_date}{Start day.}

\item{is_rdiff}{Boolean, apply Quandl percentage difference transformation.}
}
\value{
A time series
}
\description{
Get unmanaged series from Quanl.
}

