% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/search-series.R
\name{search_series}
\alias{search_series}
\title{Search series name from Quandl.}
\usage{
search_series(keyword, page = 1, source = NULL)
}
\arguments{
\item{keyword}{String}

\item{page}{Integer, number of pages}

\item{source}{String, the source filtering}
}
\description{
Search series name from Quandl.
}

